+++ priv/repo/migrations/20211115103758_fill_vf_satisfaction.exs
@@ -23,9 +23,9 @@ def change() do
 		add :satisfies_id, references("vf_intent"), null: false
 		add :satisfied_by_id, references("vf_event_or_commitment"), null: false
 		add :resource_quantity_has_unit_id, references("vf_unit")
-		add :resource_quantity_has_numerical_value, :float
+		add :resource_quantity_has_numerical_value, :decimal
 		add :effort_quantity_has_unit_id, references("vf_unit")
-		add :effort_quantity_has_numerical_value, :float
+		add :effort_quantity_has_numerical_value, :decimal
 		add :note, :text
 		timestamps()
 	end
